Prevent 'wrong_thread_help_topic_pygame' help post alerts from auto self-deleting in 'helpforums_pre' extension

+++ pcbot/exts/helpforums_pre/cogs.py
@@ -1513,24 +1513,25 @@ class HelpForumsPreCog(BaseExtensionCog, name="helpforums-pre"):
             embed=discord.Embed(
                 title="Your post is about pygame(-ce). We have better channels for that!",
                 description=(
-                    "Your post is about pygame(-ce). We have better channels for that!\n\n"
-                    "**Please delete your post and recreate it in one of these channels, "
-                    "based on your roles:**\n\n"
+                    "**Please delete your starter message and recreate your post in one of "
+                    "these channels, based on your roles:**\n\n"
                     f"- <#{HELP_FORUM_CHANNEL_IDS['newbies']}>\n"
-                    f"- <#{HELP_FORUM_CHANNEL_IDS['regulars']}>\n"
+                    f"- <#{HELP_FORUM_CHANNEL_IDS['regulars']}>\n\n"
                     "**Thank you for helping us maintain clean help forum channels** "
                     "<:pg_robot:837389387024957440>\n\n"
                     "This alert should disappear after you've deleted your post.\n\n"
                     "Did I get it wrong? If yes, please react with ☝️ to dismiss "
-                    f"this alert <t:{int(time.time() + 60)}:R>. *Don't lie* ;)"
+                    f"this alert <t:{int(time.time() + 60)}:R>. *Don't lie* 😉"
                 ),
                 color=0x36393F,
             ),
         )
         await message.add_reaction("☝️")
 
+        event = None
+
         try:
-            await self.bot.wait_for(
+            event = await self.bot.wait_for(
                 "raw_reaction_add",
                 check=lambda event: event.message_id == message.id
                 and (
@@ -1552,7 +1553,8 @@ class HelpForumsPreCog(BaseExtensionCog, name="helpforums-pre"):
             )
         except asyncio.TimeoutError:
             pass
-        else:
+
+        if event:
             try:
                 await message.delete()
             except discord.HTTPException:
